Section calendar/gallery components into separate folder
* With the addition of the gallery, DataCommon is now serving enough purposes to warrant further component organization. The components for the January and February 2020 calendar were already organized in a separate folder; this commit simply moves other relevant components (such as Gallery and YearNavItem) into a similar folder structure.
