+++ app/javascript/pages/components/gallery/BlankCalendarItem.jsx
