+++ app/javascript/pages/components/gallery/YearNavItem.jsx
